+++ launchUI.bash
@@ -7,14 +7,14 @@ cd build
 if [ "$input_theme" == "classic" ]; then
       echo "classic theme selected"
       #start dark theme
-      python ../launchUI_classic.py
+      python ../mode_classic.py
 elif [ "$input_theme" == "light" ]; then
       echo "light theme selected"
       #start light theme
-      python ../launchUI.py --theme light
+      python ../mode_dark_light.py --theme light
 else
       echo "No theme selected, will start default theme (dark mode)"
       #start default theme (dark mode)
-      python ../launchUI.py
+      python ../mode_dark_light.py
 fi
